:bug: Fix partial ticks not being calculated correctly

+++ Galton-Brett Simulation/src/tk/alexander01998/galton_brett/simulation/Simulation.java
@@ -74,7 +74,7 @@ public class Simulation
 			{
 				frameStart = System.nanoTime();
 				
-				float partialTicks = cycleStart - lastTick / 50;
+				float partialTicks = (float)(cycleStart - lastTick) / 50F;
 				GaltonBrett.frame.simulationPanel.render(partialTicks);
 				
 				lastFrame = cycleStart;
